src/data/usecases/survey-result/load-survey-result/db-load-survey-result.ts
import { LoadSurveyResultRepository } from '@/data/protocols/db/survey-result/load-survey-result-repository'
import { SurveyResultModel } from '@/domain/models/survey-result'
import { LoadSurveyResult } from '@/domain/usecases/survey-result/load-survey-result'
import { LoadSurveyByIdRepository } from '../../survey/load-survey-by-id/db-load-survey-by-id-protocols'

export class DbLoadSurveyResult implements LoadSurveyResult {
  constructor (
    private readonly loadSurveyResultRepository: LoadSurveyResultRepository,
    private readonly loadSurveyByIdRepository: LoadSurveyByIdRepository
  ) {}

  async load (surveyId: string): Promise<SurveyResultModel> {
    const loadSurveyResult = await this.loadSurveyResultRepository.loadResultBySurveyId(surveyId)
    if (!loadSurveyResult) {
      await this.loadSurveyByIdRepository.loadById(surveyId)
      return null
    }
    return loadSurveyResult
  }
}
